+++ R/bq-download.R
@@ -104,16 +104,8 @@ bq_table_download <-
       message("Downloading first chunk of data.")
     }
 
-    if (is.null(page_size)) {
-      chunk_size_from_user <- FALSE
-    } else {
-      assert_that(
-        is.numeric(page_size),
-        length(page_size) == 1,
-        page_size > 0
-      )
-      chunk_size_from_user <- TRUE
-    }
+    check_number_whole(page_size, min = 0, allow_null = TRUE)
+    chunk_size_from_user <- !is.null(page_size)
     chunk_size <- page_size
 
     chunk_plan <- bq_download_plan(
@@ -147,10 +139,9 @@ bq_table_download <-
     }
 
     if (chunk_size_from_user && n_got < chunk_size) {
-      abort(c(
+      cli::cli_abort(c(
         "First chunk is incomplete:",
-        x = glue("{big_mark(chunk_size)} rows were requested, but only \\
-                  {big_mark(n_got)} rows were received."),
+        x = "{big_mark(chunk_size)} rows were requested, but only {big_mark(n_got)} rows were received.",
         i = "Leave `page_size` unspecified or use an even smaller value."
       ))
     }
@@ -158,7 +149,7 @@ bq_table_download <-
     # break rest of work into natural chunks ----
     if (!chunk_size_from_user) {
       if (!bq_quiet(quiet)) {
-        message(glue("Received {big_mark(n_got)} rows in the first chunk."))
+        cli::cli_inform("Received {big_mark(n_got)} rows in the first chunk.")
       }
       chunk_size <- trunc(0.75 * n_got)
     }
@@ -178,11 +169,10 @@ bq_table_download <-
     )
 
     if (!bq_quiet(quiet)) {
-      message(glue_data(
-        chunk_plan,
-        "Downloading the remaining {big_mark(n_max)} rows in {n_chunks} \\
-         chunks of (up to) {big_mark(chunk_size)} rows."
-      ))
+      cli::cli_inform(
+        "Downloading the remaining {big_mark(chunk_plan$n_max)} rows in {chunk_plan$n_chunks} \\
+         chunks of (up to) {big_mark(chunk_plan$chunk_size)} rows."
+      )
     }
 
     for (i in seq_len(chunk_plan$n_chunks)) {
@@ -237,8 +227,8 @@ rapply_int64 <- function(x, f) {
 }
 
 set_row_params <- function(nrow, n_max = Inf, start_index = 0L) {
-  assert_that(is.numeric(n_max), length(n_max) == 1, n_max >= 0)
-  assert_that(is.numeric(start_index), length(start_index) == 1, start_index >= 0)
+  check_number_whole(n_max, min = 0, allow_infinite = TRUE)
+  check_number_whole(start_index, min = 0)
 
   n_max <- max(min(n_max, nrow - start_index), 0)
 
@@ -288,8 +278,8 @@ set_chunk_plan <- function(n_max, chunk_size, n_chunks, start_index = 0) {
 
 bq_download_chunk_handle <- function(x, begin = 0L, max_results = 1e4) {
   x <- as_bq_table(x)
-  assert_that(is.numeric(begin), length(begin) == 1)
-  assert_that(is.numeric(max_results), length(max_results) == 1)
+  check_number_whole(begin, min = 0)
+  check_number_whole(max_results, min = 1, allow_infinite = TRUE)
 
   # Pre-format query params with forced non-scientific notation, since the BQ
   # API doesn't accept numbers like 1e5. See issue #395 for details.