+++ R/utils.R
@@ -31,18 +31,10 @@ bq_progress <- function(..., quiet = NA) {
 }
 
 bq_check_namespace <- function(pkg, bq_type) {
-  if (requireNamespace(pkg, quietly = TRUE)) {
-    return()
-  }
-
-  abort(glue(
-    "Package '{pkg}' must be installed to load BigQuery field with type '{bq_type}'"
-  ))
+  check_installed(pkg, sprintf("to parse BigQuery '%s' fields.", bq_type))
 }
 
 isFALSE <- function(x) identical(x, FALSE)
-
-is_string <- function(x) length(x) == 1L && is.character(x)
 
 cat_line <- function(...) {
   cat(paste0(..., "\n", collapse = ""))