+++ NEWS.md
@@ -3,6 +3,9 @@
 * If `bq_job_wait()` receives a 503 response, it now waits for 2 seconds and
   tries again (#535).
 
+* `grepl(pattern, x)` is now correctly translated to 
+  `REGEXP_CONTAINS(x, pattern)` (#416).
+
 * `median()` gets a translation that works in `summarise()` and a clear
   error if you use it in `mutate()` (#419).